refactor: update project structure and refine styles for light and dark themes
