+++ src/Layout/Footer/Footer.module.scss
@@ -4,6 +4,6 @@
   align-items: center;
   position: sticky;
   bottom: 0px;
-  background-color: gray;
+  background-color: var(--bac-color-secondary-h);
   border-radius: 12px 12px 0 0;
 }
