+++ src/Layout/Footer/components/SelectedControl/SelectedControl.module.scss
@@ -10,14 +10,16 @@
   }
 
   button, a {
-    padding: 5px 10px;
+    padding: 10px 15px;
     border-radius: 8px;
+    font-size: 14px;
+    line-height: 18px;
     outline: none;
     border: none;
     display: block;
     text-decoration: none;
-    background: purple;
-    color: aqua;
+    background: var(--bac-color-card);
+    color: var(--color-primary);
     cursor: pointer;
   }
 }
