+++ src/Layout/Header/Header.module.scss
@@ -1,27 +1,15 @@
 .header {
-  padding: 40px 0;
-  width: 100%;
-  margin: 0 auto;
-  max-width: 1280px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 10px 20px;
 
-  .search {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    flex-wrap: wrap;
-    gap: 25px;
+  background-color: var(--bac-color-card);
 
-    input {
-      width: 100%;
-      max-width: 300px;
-      padding: 10px 15px;
-      border-radius: 10px;
-    }
-
-    button {
-      padding: 10px 20px;
-      border-radius: 10px;
-      background-color: yellow;
-    }
+  h2 {
+    font-family: "Barrio", serif;
+    letter-spacing: 0.05em;
+    font-size: 36px;
+    color: var(--color-primary);
   }
 }
