+++ src/Layout/Layout/Layout.module.scss
@@ -1,0 +1,9 @@
+.layout {
+  display: flex;
+  flex-direction: column;
+  gap: 32px;
+
+  &.dark {
+    cursor: default;
+  }
+}
