+++ src/Layout/Layout/Layout.tsx
@@ -1,0 +1,20 @@
+import styles from './Layout.module.scss';
+import { Outlet } from 'react-router-dom';
+import Header from '../Header/Header';
+import Footer from '../Footer/Footer';
+import { useTheme } from '@/providers/themeContext/useTheme';
+
+const Layout = () => {
+  const { theme } = useTheme();
+  return (
+    <div className={`${styles.layout} ${theme === 'dark' ? 'dark' : 'light'}`}>
+      <Header />
+      <main className={styles.main}>
+        <Outlet />
+      </main>
+      <Footer />
+    </div>
+  );
+};
+
+export default Layout;
