--- src/Layout/Main/Main.module.scss
@@ -1,5 +1,0 @@
-.main {
-  margin: 0 auto;
-  display: flex;
-  gap: 48px;
-}
