+++ src/pages/Details/Details.module.scss
@@ -10,6 +10,7 @@
   padding: 12px;
   border-radius: 24px;
   position: relative;
+    color: var(--color-primary);
 
   h2 {
     text-align: center;
@@ -18,6 +19,8 @@
     align-self: center;
     padding: 10px 15px;
     border-radius: 12px;
+    background-color: var(--bac-color-card-h);
+    color: var(--color-primary);
   }
   .wrapper {
     position: sticky;
