+++ src/pages/Home/Home.module.scss
@@ -1,14 +1,13 @@
-.home {
-  padding: 0px;
+.wrapper {
+  margin: 0 auto;
   display: flex;
-  flex-direction: column;
+  gap: 48px;
+  backdrop-filter: blur(10px);
 
-  gap: 32px;
-
-  &__dark {
-    background-color: black;
-  }
-  &__light {
-    background-color: #fafafa;
+  .main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    margin: 0 auto;
   }
 }
