+++ src/pages/Home/modules/Card/constants.ts
