+++ src/pages/Home/modules/Card/types.ts
