+++ src/pages/Home/modules/CardList/types.ts
