+++ src/pages/Home/modules/NoResult/NoResult.module.scss
