+++ src/pages/Home/modules/NoResult/constants.ts
