+++ src/pages/Home/modules/NoResult/types.ts
