+++ src/pages/Home/modules/Pagination/Pagination.module.scss
@@ -10,12 +10,15 @@
     display: flex;
     justify-content: center;
     align-items: center;
-    color: white;
-    background-color: #000;
+    background-color: var(--bac-color-card-h);
+    color: var(--color-primary);
 
     &:disabled {
       color: gray;
     background-color: gray;
     }
+  }
+  span {
+    color: var(--color-primary);
   }
 }
