+++ src/pages/Home/modules/Results/Results.module.scss
